refactor: replace fmt messages with log
