+++ main.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/bwmarrin/discordgo"
 	"github.com/mmm-development/bulker-discord/appcmd"
+	"github.com/mmm-development/bulker-discord/clog"
 )
 
 var (
@@ -17,14 +18,14 @@ var (
 
 func init() {
 	flag.Parse()
+
+	clog.L.Register(os.Stdout, clog.INFO)
 }
 
 func main() {
 	dg, err := discordgo.New("Bot " + *Token)
 	if err != nil {
-		fmt.Println("[ERROR] Creating Discord session:")
-		fmt.Println(err)
-		return
+		clog.L.Fatal(fmt.Sprintf("Creating Discord Session:\n%v", err))
 	}
 
 	dg.AddHandler(ready)
@@ -34,43 +35,37 @@ func main() {
 
 	err = dg.Open()
 	if err != nil {
-		fmt.Println("[ERROR] Opening connection:")
-		fmt.Println(err)
-		return
+		clog.L.Fatal(fmt.Sprintf("Opening connection:\n%v", err))
 	}
 	defer dg.Close()
 
-	fmt.Println("[INFO] Registering commands...")
+	clog.L.Info("Registering commands...")
 	registeredCommands := make([]*discordgo.ApplicationCommand, len(appcmd.Commands))
 	for i, v := range appcmd.Commands {
 		cmd, err := dg.ApplicationCommandCreate(dg.State.User.ID, "", v)
 		if err != nil {
-			fmt.Printf("[ERROR] Creating '%v' command:\n", v.Name)
-			fmt.Println(err)
-			return
+			clog.L.Fatal(fmt.Sprintf("Creating '%v' command:\n%v", v.Name, err))
 		}
 		registeredCommands[i] = cmd
 	}
 	defer func() {
-		fmt.Println("[INFO] Cleaning up registered commands...")
+		clog.L.Info("Cleaning up registered commands...")
 		for _, v := range registeredCommands {
 			err := dg.ApplicationCommandDelete(dg.State.User.ID, "", v.ID)
 			if err != nil {
-				fmt.Printf("[ERROR] Removing '%v' command:\n", v.Name)
-				fmt.Println(err)
-				return
+				clog.L.Fatal(fmt.Sprintf("Removing '%v' command:\n%v", v.Name, err))
 			}
 		}
 	}()
 
-	fmt.Println("[INFO] Bot is running, press CTRL-C to exit")
+	clog.L.Info("Bot is running, press CTRL-C to exit")
 	stop := make(chan os.Signal, 1)
 	signal.Notify(stop, syscall.SIGINT, syscall.SIGTERM, os.Interrupt)
 	<-stop
 }
 
 func ready(s *discordgo.Session, r *discordgo.Ready) {
-	fmt.Printf("[INFO] Bot %s is ready\n", r.User.Username)
+	clog.L.Info(fmt.Sprintf("Bot %s is ready", r.User.Username))
 }
 
 func interactionCreate(s *discordgo.Session, i *discordgo.InteractionCreate) {
